soa-login/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
supabase==2.24.0
PyJWT==2.10.1
cryptography==46.0.3
email-validator==2.3.0
requests==2.32.3
pika==1.3.2
